requirements.txt
torch
torchvision
pillow
matplotlib
seaborn
pandas
scikit-learn
